Generalized tokenpicker for quote selection
