+++ packages/ui-library/src/components/index.tsx
@@ -77,3 +77,4 @@ export * from './background-paper';
 export * from './select';
 export * from './container-box';
 export * from './token-amount-usd-input';
+export * from './token-picker-button';
